refactor: consolidate admin resource command helpers
