+++ cmd/admin/dataserver/cli/option.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package option
+package cli
 
 import "github.com/spf13/pflag"
