+++ cmd/admin/dataserver/create/cmd_test.go
@@ -26,7 +26,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/oxia-db/oxia/cmd/admin/commons"
-	"github.com/oxia-db/oxia/cmd/admin/dataserver/option"
+	dataservercli "github.com/oxia-db/oxia/cmd/admin/dataserver/cli"
 	"github.com/oxia-db/oxia/common/proto"
 )
 
@@ -76,8 +76,8 @@ func Test_cmd_createDataServer(t *testing.T) {
 		SilenceUsage: Cmd.SilenceUsage,
 	}
 	fields.AddFlags(cmd.Flags())
-	_ = cmd.MarkFlagRequired(option.PublicFlagName)
-	_ = cmd.MarkFlagRequired(option.InternalFlagName)
+	_ = cmd.MarkFlagRequired(dataservercli.PublicFlagName)
+	_ = cmd.MarkFlagRequired(dataservercli.InternalFlagName)
 	out, err := runCmd(cmd, serverName, "--public", "public1", "--internal", "internal1", "--label", "rack=rack-1", "-o", "json")
 
 	assert.NoError(t, err)
@@ -117,8 +117,8 @@ func Test_cmd_createDataServer_DefaultTable(t *testing.T) {
 		SilenceUsage: Cmd.SilenceUsage,
 	}
 	fields.AddFlags(cmd.Flags())
-	_ = cmd.MarkFlagRequired(option.PublicFlagName)
-	_ = cmd.MarkFlagRequired(option.InternalFlagName)
+	_ = cmd.MarkFlagRequired(dataservercli.PublicFlagName)
+	_ = cmd.MarkFlagRequired(dataservercli.InternalFlagName)
 	out, err := runCmd(cmd, serverName, "--public", "public1", "--internal", "internal1")
 
 	assert.NoError(t, err)
@@ -154,8 +154,8 @@ func Test_cmd_createDataServer_Name(t *testing.T) {
 		SilenceUsage: Cmd.SilenceUsage,
 	}
 	fields.AddFlags(cmd.Flags())
-	_ = cmd.MarkFlagRequired(option.PublicFlagName)
-	_ = cmd.MarkFlagRequired(option.InternalFlagName)
+	_ = cmd.MarkFlagRequired(dataservercli.PublicFlagName)
+	_ = cmd.MarkFlagRequired(dataservercli.InternalFlagName)
 	out, err := runCmd(cmd, serverName, "--public", "public1", "--internal", "internal1", "-o", "name")
 
 	assert.NoError(t, err)
@@ -175,8 +175,8 @@ func Test_cmd_createDataServer_InvalidLabel(t *testing.T) {
 		SilenceUsage: Cmd.SilenceUsage,
 	}
 	fields.AddFlags(cmd.Flags())
-	_ = cmd.MarkFlagRequired(option.PublicFlagName)
-	_ = cmd.MarkFlagRequired(option.InternalFlagName)
+	_ = cmd.MarkFlagRequired(dataservercli.PublicFlagName)
+	_ = cmd.MarkFlagRequired(dataservercli.InternalFlagName)
 	out, err := runCmd(cmd, "server-1", "--public", "public1", "--internal", "internal1", "--label", "rack")
 
 	assert.Error(t, err)
@@ -197,8 +197,8 @@ func Test_cmd_createDataServer_RejectsEmptyName(t *testing.T) {
 		SilenceUsage: Cmd.SilenceUsage,
 	}
 	fields.AddFlags(cmd.Flags())
-	_ = cmd.MarkFlagRequired(option.PublicFlagName)
-	_ = cmd.MarkFlagRequired(option.InternalFlagName)
+	_ = cmd.MarkFlagRequired(dataservercli.PublicFlagName)
+	_ = cmd.MarkFlagRequired(dataservercli.InternalFlagName)
 	out, err := runCmd(cmd, "   ", "--public", "public1", "--internal", "internal1")
 
 	assert.Error(t, err)
